+++ core/test/unit/config_spec.js
@@ -4,6 +4,7 @@ var should         = require('should'),
     sinon          = require('sinon'),
     Promise        = require('bluebird'),
     path           = require('path'),
+    fs             = require('fs'),
     _              = require('lodash'),
     rewire         = require('rewire'),
 
@@ -12,8 +13,7 @@ var should         = require('should'),
     // Thing we are testing
     defaultConfig  = require('../../../config.example')[process.env.NODE_ENV],
     theme          = rewire('../../server/config/theme'),
-    config         = rewire('../../server/config'),
-    configUpdate   = config.__get__('updateConfig');
+    config         = rewire('../../server/config');
 
 // To stop jshint complaining
 should.equal(true, true);
@@ -76,7 +76,7 @@ describe('Config', function () {
             // Make a copy of the default config file
             // so we can restore it after every test.
             // Using _.merge to recursively apply every property.
-            configUpdate(_.merge({}, config));
+            config.set(_.merge({}, config));
         });
 
         it('should have exactly the right keys', function () {
@@ -114,24 +114,24 @@ describe('Config', function () {
         });
 
         it('should not return a slash for subdir', function () {
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.paths.should.have.property('subdir', '');
 
-            configUpdate({url: 'http://my-ghost-blog.com/'});
+            config.set({url: 'http://my-ghost-blog.com/'});
             config.paths.should.have.property('subdir', '');
         });
 
         it('should handle subdirectories properly', function () {
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.paths.should.have.property('subdir', '/blog');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog/'});
+            config.set({url: 'http://my-ghost-blog.com/blog/'});
             config.paths.should.have.property('subdir', '/blog');
 
-            configUpdate({url: 'http://my-ghost-blog.com/my/blog'});
+            config.set({url: 'http://my-ghost-blog.com/my/blog'});
             config.paths.should.have.property('subdir', '/my/blog');
 
-            configUpdate({url: 'http://my-ghost-blog.com/my/blog/'});
+            config.set({url: 'http://my-ghost-blog.com/my/blog/'});
             config.paths.should.have.property('subdir', '/my/blog');
         });
 
@@ -139,7 +139,7 @@ describe('Config', function () {
             var contentPath = path.join(config.paths.appRoot, 'otherContent', '/'),
                 configFile = 'configFileDanceParty.js';
 
-            configUpdate({
+            config.set({
                 config: configFile,
                 paths: {
                     contentPath: contentPath
@@ -157,27 +157,27 @@ describe('Config', function () {
     describe('urlFor', function () {
 
         before(function () {
-            configUpdate(_.merge({}, defaultConfig));
+            config.set(_.merge({}, defaultConfig));
         });
 
         afterEach(function () {
-            configUpdate({url: defaultConfig.url});
+            config.set({url: defaultConfig.url});
         });
 
         it('should return the home url with no options', function () {
             config.urlFor().should.equal('/');
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor().should.equal('/blog/');
         });
 
         it('should return home url when asked for', function () {
             var testContext = 'home';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext).should.equal('/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext).should.equal('/blog/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/blog/');
         });
@@ -185,11 +185,11 @@ describe('Config', function () {
         it('should return rss url when asked for', function () {
             var testContext = 'rss';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext).should.equal('/rss/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/rss/');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext).should.equal('/blog/rss/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/blog/rss/');
         });
@@ -197,11 +197,11 @@ describe('Config', function () {
         it('should return url for a random path when asked for', function () {
             var testContext = {relativeUrl: '/about/'};
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext).should.equal('/about/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/about/');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext).should.equal('/blog/about/');
             config.urlFor(testContext, true).should.equal('http://my-ghost-blog.com/blog/about/');
         });
@@ -210,11 +210,11 @@ describe('Config', function () {
             var testContext = 'post',
                 testData = {post: testUtils.DataGenerator.Content.posts[2], permalinks: {value: '/:slug/'}};
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext, testData).should.equal('/short-and-sweet/');
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com/short-and-sweet/');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext, testData).should.equal('/blog/short-and-sweet/');
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com/blog/short-and-sweet/');
         });
@@ -231,11 +231,11 @@ describe('Config', function () {
                 yyyy = today.getFullYear(),
                 postLink = '/' + yyyy + '/' + mm + '/' + dd + '/short-and-sweet/';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext, testData).should.equal(postLink);
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com' + postLink);
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext, testData).should.equal('/blog' + postLink);
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com/blog' + postLink);
         });
@@ -244,11 +244,11 @@ describe('Config', function () {
             var testContext = 'tag',
                 testData = {tag: testUtils.DataGenerator.Content.tags[0]};
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
             config.urlFor(testContext, testData).should.equal('/tag/kitchen-sink/');
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com/tag/kitchen-sink/');
 
-            configUpdate({url: 'http://my-ghost-blog.com/blog'});
+            config.set({url: 'http://my-ghost-blog.com/blog'});
             config.urlFor(testContext, testData).should.equal('/blog/tag/kitchen-sink/');
             config.urlFor(testContext, testData, true).should.equal('http://my-ghost-blog.com/blog/tag/kitchen-sink/');
         });
@@ -264,7 +264,7 @@ describe('Config', function () {
 
         afterEach(function () {
             sandbox.restore();
-            configUpdate({url: defaultConfig.url});
+            config.set({url: defaultConfig.url});
         });
 
         it('should output correct url for post', function (done) {
@@ -276,7 +276,7 @@ describe('Config', function () {
                 testData = testUtils.DataGenerator.Content.posts[2],
                 postLink = '/short-and-sweet/';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
 
             // next test
             config.urlForPost(settings, testData).then(function (url) {
@@ -287,7 +287,7 @@ describe('Config', function () {
             }).then(function (url) {
                 url.should.equal('http://my-ghost-blog.com' + postLink);
 
-                return configUpdate({url: 'http://my-ghost-blog.com/blog'});
+                return config.set({url: 'http://my-ghost-blog.com/blog'});
             }).then(function () {
 
                 // next test
@@ -318,7 +318,7 @@ describe('Config', function () {
                 yyyy = today.getFullYear(),
                 postLink = '/' + yyyy + '/' + mm + '/' + dd + '/short-and-sweet/';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
 
             // next test
             config.urlForPost(settings, testData).then(function (url) {
@@ -329,7 +329,7 @@ describe('Config', function () {
             }).then(function (url) {
                 url.should.equal('http://my-ghost-blog.com' + postLink);
 
-                return configUpdate({url: 'http://my-ghost-blog.com/blog'});
+                return config.set({url: 'http://my-ghost-blog.com/blog'});
             }).then(function () {
 
                 // next test
@@ -355,7 +355,7 @@ describe('Config', function () {
                 testData = testUtils.DataGenerator.Content.posts[5],
                 postLink = '/static-page-test/';
 
-            configUpdate({url: 'http://my-ghost-blog.com'});
+            config.set({url: 'http://my-ghost-blog.com'});
 
             // next test
             config.urlForPost(settings, testData).then(function (url) {
@@ -366,7 +366,7 @@ describe('Config', function () {
             }).then(function (url) {
                 url.should.equal('http://my-ghost-blog.com' + postLink);
 
-                return configUpdate({url: 'http://my-ghost-blog.com/blog'});
+                return config.set({url: 'http://my-ghost-blog.com/blog'});
             }).then(function () {
 
                 // next test
@@ -378,6 +378,329 @@ describe('Config', function () {
                 return config.urlForPost(settings, testData, true);
             }).then(function (url) {
                 url.should.equal('http://my-ghost-blog.com/blog' + postLink);
+
+                done();
+            }).catch(done);
+        });
+    });
+
+    describe('File', function () {
+        var sandbox,
+            originalConfig,
+            readFileStub,
+            overrideConfig = function (newConfig) {
+                readFileStub.returns(
+                    _.extend({}, defaultConfig, newConfig)
+                );
+            },
+            expectedError = new Error('expected bootstrap() to throw error but none thrown');
+
+        before(function () {
+            originalConfig = _.cloneDeep(rewire('../../server/config')._config);
+        });
+
+        beforeEach(function () {
+            sandbox = sinon.sandbox.create();
+            readFileStub = sandbox.stub(config, 'readFile');
+        });
+
+        afterEach(function () {
+            config = rewire('../../server/config');
+            sandbox.restore();
+        });
+
+        it('loads the config file if one exists', function (done) {
+            // We actually want the real method here.
+            readFileStub.restore();
+
+            // the test infrastructure is setup so that there is always config present,
+            // but we want to overwrite the test to actually load config.example.js, so that any local changes
+            // don't break the tests
+            config.set({
+                paths: {
+                    appRoot: path.join(originalConfig.paths.appRoot, 'config.example.js')
+                }
+            });
+
+            config.load().then(function (config) {
+                config.url.should.equal(defaultConfig.url);
+                config.database.client.should.equal(defaultConfig.database.client);
+                config.database.connection.should.eql(defaultConfig.database.connection);
+                config.server.host.should.equal(defaultConfig.server.host);
+                config.server.port.should.equal(defaultConfig.server.port);
+
+                done();
+            }).catch(done);
+        });
+
+        it('uses the passed in config file location', function (done) {
+            // We actually want the real method here.
+            readFileStub.restore();
+
+            config.load(path.join(originalConfig.paths.appRoot, 'config.example.js')).then(function (config) {
+                config.url.should.equal(defaultConfig.url);
+                config.database.client.should.equal(defaultConfig.database.client);
+                config.database.connection.should.eql(defaultConfig.database.connection);
+                config.server.host.should.equal(defaultConfig.server.host);
+                config.server.port.should.equal(defaultConfig.server.port);
+
+                done();
+            }).catch(done);
+        });
+
+        it('creates the config file if one does not exist', function (done) {
+                // trick bootstrap into thinking that the config file doesn't exist yet
+            var existsStub = sandbox.stub(fs, 'exists', function (file, cb) { return cb(false); }),
+                // ensure that the file creation is a stub, the tests shouldn't really create a file
+                writeFileStub = sandbox.stub(config, 'writeFile').returns(Promise.resolve()),
+                validateStub = sandbox.stub(config, 'validate').returns(Promise.resolve());
+
+            config.load().then(function () {
+                existsStub.calledOnce.should.be.true;
+                writeFileStub.calledOnce.should.be.true;
+                validateStub.calledOnce.should.be.true;
+                done();
+            }).catch(done);
+        });
+
+        it('accepts urls with a valid scheme', function (done) {
+            // replace the config file with invalid data
+            overrideConfig({url: 'http://testurl.com'});
+
+            config.load().then(function (localConfig) {
+                localConfig.url.should.equal('http://testurl.com');
+
+                // Next test
+                overrideConfig({url: 'https://testurl.com'});
+                return config.load();
+            }).then(function (localConfig) {
+                localConfig.url.should.equal('https://testurl.com');
+
+                 // Next test
+                overrideConfig({url: 'http://testurl.com/blog/'});
+                return config.load();
+            }).then(function (localConfig) {
+                localConfig.url.should.equal('http://testurl.com/blog/');
+
+                 // Next test
+                overrideConfig({url: 'http://testurl.com/ghostly/'});
+                return config.load();
+            }).then(function (localConfig) {
+                localConfig.url.should.equal('http://testurl.com/ghostly/');
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects a fqdn without a scheme', function (done) {
+
+            overrideConfig({ url: 'example.com' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects a hostname without a scheme', function (done) {
+
+            overrideConfig({ url: 'example' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects a hostname with a scheme', function (done) {
+
+            overrideConfig({ url: 'https://example' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects a url with an unsupported scheme', function (done) {
+
+            overrideConfig({ url: 'ftp://example.com' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects a url with a protocol relative scheme', function (done) {
+
+            overrideConfig({ url: '//example.com' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('does not permit the word ghost as a url path', function (done) {
+            overrideConfig({ url: 'http://example.com/ghost/' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('does not permit the word ghost to be a component in a url path', function (done) {
+            overrideConfig({ url: 'http://example.com/blog/ghost/' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('does not permit the word ghost to be a component in a url path', function (done) {
+            overrideConfig({ url: 'http://example.com/ghost/blog/' });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('does not permit database config to be falsy', function (done) {
+            // replace the config file with invalid data
+            overrideConfig({ database: false });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('does not permit database config to be empty', function (done) {
+            // replace the config file with invalid data
+            overrideConfig({ database: {} });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+
+        it('requires server to be present', function (done) {
+            overrideConfig({ server: false });
+
+            config.load().then(function (localConfig) {
+                /*jshint unused:false*/
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('allows server to use a socket', function (done) {
+            overrideConfig({ server: { socket: 'test' } });
+
+            config.load().then(function (localConfig) {
+                should.exist(localConfig);
+                localConfig.server.socket.should.equal('test');
+
+                done();
+            }).catch(done);
+        });
+
+        it('allows server to have a host and a port', function (done) {
+            overrideConfig({ server: { host: '127.0.0.1', port: '2368' } });
+
+            config.load().then(function (localConfig) {
+                should.exist(localConfig);
+                localConfig.server.host.should.equal('127.0.0.1');
+                localConfig.server.port.should.equal('2368');
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects server if there is a host but no port', function (done) {
+            overrideConfig({ server: { host: '127.0.0.1' } });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects server if there is a port but no host', function (done) {
+            overrideConfig({ server: { port: '2368' } });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
+
+                done();
+            }).catch(done);
+        });
+
+        it('rejects server if configuration is empty', function (done) {
+            overrideConfig({ server: {} });
+
+            config.load().then(function () {
+                done(expectedError);
+            }).catch(function (err) {
+                should.exist(err);
+                err.should.be.an.Error;
 
                 done();
             }).catch(done);
